+++ src/wajuce.c
@@ -16,7 +16,9 @@ static int32_t next_id = 1;
 // ============================================================================
 
 FFI_PLUGIN_EXPORT int32_t wajuce_context_create(int32_t sample_rate,
-                                                int32_t buffer_size) {
+                                                int32_t buffer_size,
+                                                int32_t input_channels,
+                                                int32_t output_channels) {
   return next_id++;
 }
 
@@ -26,6 +28,22 @@ FFI_PLUGIN_EXPORT double wajuce_context_get_time(int32_t ctx_id) { return 0.0; }
 
 FFI_PLUGIN_EXPORT double wajuce_context_get_sample_rate(int32_t ctx_id) {
   return 44100.0;
+}
+
+FFI_PLUGIN_EXPORT int32_t wajuce_context_get_bit_depth(int32_t ctx_id) {
+  return 32;
+}
+
+FFI_PLUGIN_EXPORT int32_t
+wajuce_context_set_preferred_sample_rate(int32_t ctx_id,
+                                         double preferred_sample_rate) {
+  return 0;
+}
+
+FFI_PLUGIN_EXPORT int32_t
+wajuce_context_set_preferred_bit_depth(int32_t ctx_id,
+                                       int32_t preferred_bit_depth) {
+  return 0;
 }
 
 FFI_PLUGIN_EXPORT int32_t wajuce_context_get_state(int32_t ctx_id) { return 0; }